SRH Insert/Injection For Greater Good (#25)
* SRH Insert/Injection For Greater Good

* Insert/Inject SRH into Ipv6 Header
  ** unit and pcap/integration test srv6-inject
  ** safe writes with genericarray when creating SRHs, but keep unsafe reads otherwise via the structure memory layout
* cleanup some check.sh test script stuff to actually blow up on failure
* u32/u128 addresses are now Ipv4Addr and Ipv6Addr respectively
* colorized print examples
* macros are good for you and your compiler homies, in order to get safe/compile-time writes when creating an srh
* insert_header w/ Result type and cleanup on it and push_header code

+++ framework/tests/address.rs
@@ -5,7 +5,7 @@ use std::str::FromStr;
 
 #[test]
 fn address_inline() {
-    let pfx = Ipv4Prefix::new(u32::from(Ipv4Addr::from_str("192.168.0.0").unwrap()), 16);
+    let pfx = Ipv4Prefix::new(Ipv4Addr::from_str("192.168.0.0").unwrap(), 16);
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("192.168.0.1").unwrap()),));
     assert!(pfx.in_range(u32::from_be(16820416)));
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("192.168.100.1").unwrap()),));
@@ -14,15 +14,15 @@ fn address_inline() {
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("192.168.0.0").unwrap()),));
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("192.168.255.255").unwrap()),));
 
-    let pfx = Ipv4Prefix::new(u32::from(Ipv4Addr::from_str("192.168.1.2").unwrap()), 32);
+    let pfx = Ipv4Prefix::new(Ipv4Addr::from_str("192.168.1.2").unwrap(), 32);
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("192.168.1.2").unwrap()),));
     assert!(!pfx.in_range(u32::from(Ipv4Addr::from_str("192.168.1.3").unwrap()),));
 
-    let pfx = Ipv4Prefix::new(u32::from(Ipv4Addr::from_str("0.0.0.0").unwrap()), 0);
+    let pfx = Ipv4Prefix::new(Ipv4Addr::from_str("0.0.0.0").unwrap(), 0);
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("192.168.1.2").unwrap()),));
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("2.2.2.2").unwrap()),));
 
-    let pfx = Ipv4Prefix::new(u32::from(Ipv4Addr::from_str("0.0.0.0").unwrap()), 0);
+    let pfx = Ipv4Prefix::new(Ipv4Addr::from_str("0.0.0.0").unwrap(), 0);
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("192.168.1.2").unwrap()),));
     assert!(pfx.in_range(u32::from(Ipv4Addr::from_str("2.2.2.2").unwrap()),));
 }
